+++ winui3/SquirrelNotifier.WinUI3/Properties/AssemblyInfo.cs
@@ -1,4 +1,5 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("SquirrelNotifier.WinUI3.Tests")]
+[assembly: InternalsVisibleTo("SquirrelNotifier.HeadlessE2E")]
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]